+++ src/test/java/racingcar/racing/RacingServiceTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.racing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -6,7 +6,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.car.CarAction;
 import racingcar.car.CarDTO;
-import racingcar.racing.RacingService;
 
 import java.util.ArrayList;
 import java.util.Arrays;
