+++ src/test/java/racingcar/racing/ValidatorTest.java
@@ -1,9 +1,8 @@
-package racingcar;
+package racingcar.racing;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.racing.Validator;
 
 import java.util.Arrays;
